try to fix circular import error
