+++ src/file_mover.py
@@ -1,6 +1,6 @@
 import os
 import sys
-from src.utils import CheckCCDI, set_s3_session_client, get_logger, get_date, get_time
+from src.utils import CheckCCDI, set_s3_session_client, get_logger, get_date, get_time, calculate_object_md5sum_new
 from botocore.exceptions import (
     ClientError,
     ReadTimeoutError,
@@ -29,65 +29,6 @@ def parse_file_url_in_cds(url: str) -> tuple:
     else:
         pass
     return bucket_name, object_key
-
-
-def calculate_object_md5sum(s3_client, url) -> str:
-    """Calculate md5sum of an object using url
-    This function was modified based on https://github.com/jmwarfe/s3-md5sum/blob/main/s3-md5sum.py
-
-    Example of url:
-    s3://example-bucket/folder1/folder2/test_file.fastq.gz
-    """
-    bucket_name, object_key = parse_file_url_in_cds(url)
-    # get obejct
-    obj = s3_client.get_object(Bucket=bucket_name, Key=object_key)
-    obj_body = obj["Body"]
-
-    # Initialize MD5 hash object
-    md5_hash = hashlib.md5()
-    # Read the object in chunks and update the MD5 hash
-    for chunk in iter(lambda: obj_body.read(1024 * 1024), b""):
-        md5_hash.update(chunk)
-    return md5_hash.hexdigest()
-
-
-def calculate_object_md5sum_new(s3_client, url) -> str:
-    """Calculate md5sum of an object using url
-    This function was modified based on https://github.com/jmwarfe/s3-md5sum/blob/main/s3-md5sum.py
-    The new one reads specific byte range frm file as a chunk to avoid empty chunk
-    aws server getting time out
-
-    Example of url:
-    s3://example-bucket/folder1/folder2/test_file.fastq.gz
-    """
-    # specify a chunk size to get object
-    chunk_size = 1073741824
-    # get object size
-    bucket_name, object_key = parse_file_url_in_cds(url)
-    object_size = s3_client.get_object(Bucket=bucket_name, Key=object_key)[
-        "ContentLength"
-    ]
-
-    # object_size = s3_client.get_object_attributes(
-    #    Bucket=bucket_name, Key=object_key, ObjectAttributes=["ObjectSize"]
-    # ).get("ObjectSize")
-
-    chunk_start = 0
-    chunk_end = chunk_start + chunk_size - 1
-
-    # Initialize MD5 hash object
-    md5_hash = hashlib.md5()
-    while chunk_start <= object_size:
-        # Read specific byte range from file as a chunk. We do this because AWS server times out and sends
-        # empty chunks when streaming the entire file.
-        if body := s3_client.get_object(
-            Bucket=bucket_name, Key=object_key, Range=f"bytes={chunk_start}-{chunk_end}"
-        ).get("Body"):
-            for small_chunk in iter(lambda: body.read(1024 * 1024), b""):
-                md5_hash.update(small_chunk)
-            chunk_start += chunk_size
-            chunk_end += chunk_size
-    return md5_hash.hexdigest()
 
 
 def copy_object_parameter(url_in_cds: str, dest_bucket_path: str) -> dict:
