+++ src/utils.py
@@ -23,7 +23,6 @@ from zipfile import ZipFile
 from shutil import copy
 import json
 from botocore.exceptions import ClientError
-from src.file_mover import calculate_object_md5sum_new, parse_file_url_in_cds
 
 
 ExcelFile = TypeVar("ExcelFile")
@@ -953,6 +952,65 @@ class CheckCCDI:
         # remove any duplcates
         file_node_list_uniq = list(set(file_node_list))
         return file_node_list_uniq
+
+
+def list_to_chunks(mylist: list, chunk_len: int) -> list:
+    """Break a list into a list of chunks"""
+    chunks = [
+        mylist[i * chunk_len : (i + 1) * chunk_len]
+        for i in range((len(mylist) + chunk_len - 1) // chunk_len)
+    ]
+    return chunks
+
+
+def parse_file_url_in_cds(url: str) -> tuple:
+    parsed_url = urlparse(url)
+    bucket_name = parsed_url.netloc
+    object_key = parsed_url.path
+    if object_key[0] == "/":
+        object_key = object_key[1:]
+    else:
+        pass
+    return bucket_name, object_key
+
+
+def calculate_object_md5sum_new(s3_client, url) -> str:
+    """Calculate md5sum of an object using url
+    This function was modified based on https://github.com/jmwarfe/s3-md5sum/blob/main/s3-md5sum.py
+    The new one reads specific byte range frm file as a chunk to avoid empty chunk
+    aws server getting time out
+
+    Example of url:
+    s3://example-bucket/folder1/folder2/test_file.fastq.gz
+    """
+    # specify a chunk size to get object
+    chunk_size = 1073741824
+    # get object size
+    bucket_name, object_key = parse_file_url_in_cds(url)
+    object_size = s3_client.get_object(Bucket=bucket_name, Key=object_key)[
+        "ContentLength"
+    ]
+
+    # object_size = s3_client.get_object_attributes(
+    #    Bucket=bucket_name, Key=object_key, ObjectAttributes=["ObjectSize"]
+    # ).get("ObjectSize")
+
+    chunk_start = 0
+    chunk_end = chunk_start + chunk_size - 1
+
+    # Initialize MD5 hash object
+    md5_hash = hashlib.md5()
+    while chunk_start <= object_size:
+        # Read specific byte range from file as a chunk. We do this because AWS server times out and sends
+        # empty chunks when streaming the entire file.
+        if body := s3_client.get_object(
+            Bucket=bucket_name, Key=object_key, Range=f"bytes={chunk_start}-{chunk_end}"
+        ).get("Body"):
+            for small_chunk in iter(lambda: body.read(1024 * 1024), b""):
+                md5_hash.update(small_chunk)
+            chunk_start += chunk_size
+            chunk_end += chunk_size
+    return md5_hash.hexdigest()
 
 
 @task(
